+++ src/render/gl/gl_rlight.c
@@ -76,24 +76,24 @@ void R_RenderDlight(dlight_t *light)
     vec3_t v;
     float rad;
 
-    rad = light->radius * 0.35;
+    rad = light->radius * 0.35f;
 
     VectorSubtract(light->origin, r_origin, v);
     if (Length(v) < rad) { // view is inside the dlight
-        AddLightBlend(1, 0.5, 0, light->radius * 0.0003);
+        AddLightBlend(1, 0.5f, 0, light->radius * 0.0003f);
         return;
     }
 
     glBegin(GL_TRIANGLE_FAN);
-    glColor3f(0.2, 0.1, 0.0);
+    glColor3f(0.2f, 0.1f, 0.0f);
     for (i = 0; i < 3; i++)
         v[i] = light->origin[i] - vpn[i] * rad;
     glVertex3fv(v);
     glColor3f(0, 0, 0);
     for (i = 16; i >= 0; i--) {
-        a = i / 16.0 * M_PI * 2;
+        a = i / 16.0f * M_PI * 2;
         for (j = 0; j < 3; j++)
-            v[j] = light->origin[j] + vright[j] * cos(a) * rad + vup[j] * sin(a) * rad;
+            v[j] = light->origin[j] + vright[j] * cosf(a) * rad + vup[j] * sinf(a) * rad;
         glVertex3fv(v);
     }
     glEnd();
